Move pages into one file
